+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Views
 
 Extensions
 
+- `Range.init(lower:upper:)` and `ClosedRange.init(lower:upper:)` initializers are added
 - `UIDevice.safeAreaInsets` no longer returns an `Optional`
 
 ### [4.10.0(83)](https://github.com/VakhoKontridze/VCore/releases/tag/4.10.0) — *2023 06 01*